+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Snapshot tests for `Animation`.
 - Snapshot tests for `CollectionBadges`.
 - Snapshot tests for `Container`.
+- Snapshot tests for `DiscountBadge`.
 
 ## [3.17.2] - 2019-02-18